Start work on integrating estnltk, change some old migrations that wouldn't run on sqlite, set up default Django + pytest combo, rename old Unittests-based tests that weren't used, add migrations for Latvian, Lithuanian modeltranslations, start using spatialite with pytest to successfully mock our PostGIS situation, add MyPy integration, add Codecov, clear some TODOs
