+++ Dockerfile
@@ -7,7 +7,7 @@ MAINTAINER Lauri Elias <lauri@ajapaik.ee>
 RUN apt-get update && \
     apt-get install -y --fix-missing cmake build-essential gfortran git wget curl graphicsmagick \
     libgraphicsmagick1-dev libatlas-dev libavcodec-dev libavformat-dev libgtk2.0-dev libjpeg-dev liblapack-dev \
-    libswscale-dev pkg-config python3-dev python3-numpy software-properties-common zip
+    libswscale-dev pkg-config python3-dev python3-numpy software-properties-common zip libsqlite3-mod-spatialite
 
 WORKDIR /home/docker/ajapaik
