+++ mypy.ini
@@ -1,0 +1,8 @@
+[mypy]
+plugins =
+    mypy_django_plugin.main
+    mypy_drf_plugin.main
+strict_optional = True
+
+[mypy.plugins.django-stubs]
+django_settings_module = ajapaik.settings.local
