+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = ajapaik.settings.test
+python_files = tests.py test_*.py *_tests.py
+# addopts = --cov=ajapaik --cov-report=html
